+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Initial release.
 - Add the Doctolib/NoAsymmetricalPunditAfterActions cop.
+- Add the Doctolib/OneOperationPerMigration cop.